requirements.txt
discord.py
discord-py-slash-command
discord-components
aiohttp
asyncio
python-dotenv
pymongo
dnspython
requests
dblpy
numpy